% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_sales.R
\name{explain_sales}
\alias{explain_sales}
\title{Explain Future Sales Using Python API}
\usage{
explain_sales(
  sales,
  method = "ets",
  months_ahead = 6,
  api_url = "https://rbus-api.onrender.com/explain_sales"
)
}
\arguments{
\item{sales}{A numeric vector of historical sales values.}

\item{method}{Forecasting method supported by the API ("ets" or "linear").}

\item{months_ahead}{Number of months ahead to forecast/explain (default is 6).}

\item{api_url}{The URL of the Python API endpoint (default is
"https://rbus-api.onrender.com/explain_sales").}
}
\value{
A character string containing the explanation of the forecasted sales.
}
\description{
This function sends historical sales data to a deployed Python FastAPI endpoint
and retrieves an explanation of the forecasted sales trend.
}
\examples{
sales_data <- c(100, 200, 150, 300, 250, 400)
explain_sales(
  sales = sales_data,
  method = "ets",
  months_ahead = 6,
  api_url = "https://rbus-api.onrender.com/explain_sales"
)

}
