% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sales.R
\name{predict_sales}
\alias{predict_sales}
\title{Predict future sales using different forecasting methods}
\usage{
predict_sales(
  data,
  method = c("mean3", "moving_avg", "linear", "lm", "ets"),
  window = 6
)
}
\arguments{
\item{data}{A data.frame containing at least a numeric column 'Sales'}

\item{method}{The forecasting method to use: "mean3", "moving_avg", "linear"/"lm", or "ets"}

\item{window}{Number of months to use for moving average (only for "moving_avg" method)}
}
\value{
A numeric prediction for the next sales value
}
\description{
Predict future sales using different forecasting methods
}
\examples{
sample_data <- data.frame(
  Month = c("Jan", "Feb", "Mar", "Apr", "May", "Jun"),
  Sales = c(100, 200, 250, 300, 400, 420)
)
predict_sales(sample_data, method = "mean3")
predict_sales(sample_data, method = "moving_avg", window = 2)
predict_sales(sample_data, method = "linear")
predict_sales(sample_data, method = "lm")

}
